newest version only contains 2 classes (covid/non-covid), removed urls for other classes, uploaded newest ipynb.
